requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]
openai
python-dotenv
pdfplumber
pip freeze 
httpx